refactor(capture): Capture screen에서 login을 분리하여, useCamera hook 생성
